implement generic response and mediator pipeline

+++ LL/src/Infrastructure/Persistence/Persistence.LL/Repositories/Guilds/GuildRepository.cs
@@ -1,6 +1,4 @@
 ﻿using Application.Common.Interfaces;
-using Common.Exceptions;
-using Domain.Models.Entities.Characters;
 using Domain.Models.Guilds;
 using Microsoft.EntityFrameworkCore;
 
@@ -14,10 +12,9 @@ public class GuildRepository : IGuildRepository
         _context = context;
     }
 
-    public async Task CreateAsync(Guid ownerCharacterId, string name, CancellationToken cancellationToken)
+    public async Task<bool> CreateAsync(Guid ownerCharacterId, string name, CancellationToken cancellationToken)
     {
-        if (await _context.Guilds.AnyAsync(g => g.Name == name, cancellationToken))
-            throw new Exception("Name or tag already exists.");
+        if (await _context.Guilds.AnyAsync(g => g.Name == name, cancellationToken)) return false;
 
         var newGuild = new Guild
         {
@@ -31,73 +28,64 @@ public class GuildRepository : IGuildRepository
 
         _context.Guilds.Add(newGuild);
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task<Guild?> GetMyGuildAsync(Guid characterId, CancellationToken cancellationToken)
-    {
-        var guild = await _context.Guilds
+    public async Task<Guild?> GetMyGuildAsync(Guid characterId, CancellationToken cancellationToken) =>
+        await _context.Guilds
             .Include(g => g.Members)
                 .ThenInclude(m => m.Character)
             .Include(g => g.Invites)
                 .ThenInclude(i => i.Character)
             .SingleOrDefaultAsync(g => g.Members.Select(gm => gm.CharacterId).Contains(characterId), cancellationToken);
 
-        return guild;
-    }
-
-    public async Task<List<Guild>> GetAllGuildsAsync(CancellationToken cancellationToken)
-    {
-        return await _context.Guilds
+    public async Task<List<Guild>> GetAllGuildsAsync(CancellationToken cancellationToken) =>
+        await _context.Guilds
             .Include(g => g.Owner)
             .Include(g => g.Members)
             .ToListAsync(cancellationToken);
-    }
 
-    public async Task LeaveGuildAsync(Guid characterId, CancellationToken cancellationToken)
+    public async Task<bool> LeaveGuildAsync(Guid characterId, CancellationToken cancellationToken)
     {
         var member = await _context.GuildMembers.FirstOrDefaultAsync(gm => gm.CharacterId == characterId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(member, nameof(member), characterId);
+        if (member == null) return false;
 
         _context.GuildMembers.Remove(member);
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task DisbandGuildAsync(Guid characterId, CancellationToken cancellationToken)
+    public async Task<bool> DisbandGuildAsync(Guid characterId, CancellationToken cancellationToken)
     {
         var guild = await _context.Guilds
             .Include(g => g.Members)
             .Include(g => g.Invites)
             .FirstOrDefaultAsync(g => g.OwnerId == characterId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(guild, nameof(guild), characterId);
+        if (guild == null) return false;
 
         _context.GuildMembers.RemoveRange(guild.Members);
         _context.GuildInvites.RemoveRange(guild.Invites);
         _context.Guilds.Remove(guild);
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task<GuildMember> GetGuildMember(Guid currentCharacterId, CancellationToken cancellationToken)
-    {
-        var member = await _context.GuildMembers.FirstOrDefaultAsync(gm => gm.CharacterId == currentCharacterId, cancellationToken);
+    public async Task<GuildMember?> GetGuildMember(Guid currentCharacterId, CancellationToken cancellationToken) =>
+        await _context.GuildMembers
+            .FirstOrDefaultAsync(gm => gm.CharacterId == currentCharacterId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(member, nameof(member), currentCharacterId);
-
-        return member;
-    }
-
-    public async Task InviteAsync(Guid currentCharacterId, Guid guildId, Guid invitedCharacterId, CancellationToken cancellationToken)
+    public async Task<bool> InviteAsync(Guid currentCharacterId, Guid guildId, Guid invitedCharacterId, CancellationToken cancellationToken)
     {
         var guild = await _context.Guilds
             .Include(g => g.Members)
             .Include(g => g.Invites)
             .FirstOrDefaultAsync(g => g.Id == guildId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(guild, nameof(guild), guildId);
+        if (guild == null) return false;
 
-        if (guild.Members.Count >= guild.MaxMembers)
-            throw new InvalidOperationException("Guild is full.");
+        if (guild.Members.Count >= guild.MaxMembers) return false;
 
         guild.Invites.Add(new GuildInvite
         {
@@ -106,9 +94,10 @@ public class GuildRepository : IGuildRepository
             IsInvite = true,
         });
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task InviteCharacterByNameAsync(Guid currentCharacterId, Guid guildId, string invitedCharacterName, CancellationToken cancellationToken)
+    public async Task<bool> InviteCharacterByNameAsync(Guid currentCharacterId, Guid guildId, string invitedCharacterName, CancellationToken cancellationToken)
     {
         var guild = await _context.Guilds
             .Include(g => g.Members)
@@ -118,11 +107,9 @@ public class GuildRepository : IGuildRepository
         var invitedCharacter = await _context.Characters
             .FirstOrDefaultAsync(c => c.Name.Equals(invitedCharacterName), cancellationToken);
 
-        NotFoundException.ThrowIfNull(guild, nameof(guild), guildId);
-        NotFoundException.ThrowIfNull(invitedCharacter, nameof(invitedCharacter), invitedCharacterName);
+        if (guild == null || invitedCharacter == null) return false;
 
-        if (guild.Members.Count >= guild.MaxMembers)
-            throw new InvalidOperationException("Guild is full.");
+        if (guild.Members.Count >= guild.MaxMembers) return false;
 
         guild.Invites.Add(new GuildInvite
         {
@@ -131,17 +118,18 @@ public class GuildRepository : IGuildRepository
             IsInvite = true,
         });
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task AcceptInviteAsync(Guid characterId, Guid guildId, CancellationToken cancellationToken)
+    public async Task<bool> AcceptInviteAsync(Guid characterId, Guid guildId, CancellationToken cancellationToken)
     {
         var invite = await _context.GuildInvites
             .FirstOrDefaultAsync(i => i.GuildId == guildId && i.CharacterId == characterId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(invite, nameof(invite), guildId);
+        if (invite == null) return false;
 
         // Player can not accept an invitation to a guild that they applied for
-        if (!invite.IsInvite) throw new Exception();
+        if (!invite.IsInvite) return false;
 
         _context.GuildMembers.Add(new GuildMember
         {
@@ -152,31 +140,29 @@ public class GuildRepository : IGuildRepository
         _context.GuildInvites.Remove(invite);
 
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task<List<GuildInvite>> GetMyInvitesAsync(Guid characterId, CancellationToken cancellationToken)
-    {
-        return await _context.GuildInvites
+    public async Task<List<GuildInvite>> GetMyInvitesAsync(Guid characterId, CancellationToken cancellationToken) =>
+        await _context.GuildInvites
             .Include(gi => gi.Guild)
             .Where(gi => gi.CharacterId == characterId)
             .ToListAsync(cancellationToken);
-    }
 
-    public async Task ApplyToGuildAsync(Guid characterId, Guid guildId, CancellationToken cancellationToken)
+    public async Task<bool> ApplyToGuildAsync(Guid characterId, Guid guildId, CancellationToken cancellationToken)
     {
         var character = await _context.Characters.FindAsync([characterId], cancellationToken);
 
-        NotFoundException.ThrowIfNull(character, nameof(character), characterId);
+        if (character == null) return false;
 
         var guild = await _context.Guilds
             .Include(g => g.Members)
             .Include(g => g.Invites)
             .FirstOrDefaultAsync(g => g.Id == guildId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(guild, nameof(guild), guildId);
+        if (guild == null) return false;
 
-        if (guild.Members.Count >= guild.MaxMembers)
-            throw new InvalidOperationException("Guild is full.");
+        if (guild.Members.Count >= guild.MaxMembers) return false;
 
         guild.Invites.Add(new GuildInvite
         {
@@ -185,14 +171,15 @@ public class GuildRepository : IGuildRepository
             IsInvite = false, // This means its an application to the guild, not an invitation from the guild
         });
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task ApproveApplicationAsync(Guid guildId, Guid applicationCharacterId, CancellationToken cancellationToken)
+    public async Task<bool> ApproveApplicationAsync(Guid guildId, Guid applicationCharacterId, CancellationToken cancellationToken)
     {
         var invite = await _context.GuildInvites
             .FirstOrDefaultAsync(i => i.GuildId == guildId && i.CharacterId == applicationCharacterId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(invite, nameof(invite), guildId);
+        if (invite == null) return false;
 
         _context.GuildMembers.Add(new GuildMember
         {
@@ -203,30 +190,32 @@ public class GuildRepository : IGuildRepository
         _context.GuildInvites.Remove(invite);
 
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task RejectApplicationAsync(Guid guildId, Guid applicationCharacterId, CancellationToken cancellationToken)
+    public async Task<bool> RejectApplicationAsync(Guid guildId, Guid applicationCharacterId, CancellationToken cancellationToken)
     {
         var invite = await _context.GuildInvites
             .FirstOrDefaultAsync(i => i.GuildId == guildId && i.CharacterId == applicationCharacterId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(invite, nameof(invite), guildId);
+        if (invite == null) return false;
 
         _context.GuildInvites.Remove(invite);
 
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 
-    public async Task RejectInviteAsync(Guid characterId, Guid guildId, CancellationToken cancellationToken)
+    public async Task<bool> RejectInviteAsync(Guid characterId, Guid guildId, CancellationToken cancellationToken)
     {
-
         var invite = await _context.GuildInvites
             .FirstOrDefaultAsync(i => i.GuildId == guildId && i.CharacterId == characterId, cancellationToken);
 
-        NotFoundException.ThrowIfNull(invite, nameof(invite), guildId);
+        if (invite == null) return false;
 
         _context.GuildInvites.Remove(invite);
 
         await _context.SaveChangesAsync(cancellationToken);
+        return true;
     }
 }
